mega commit. Finished intermediate course

+++ app/Http/Controllers/ProjectController.php
@@ -2,9 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\ProjectSaved;
 use App\Http\Requests\SaveProjectRequest;
+use App\Models\Category;
 use App\Models\Project;
+use Barryvdh\Debugbar\Facades\Debugbar;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 // use Illuminate\Support\Facades\DB;
@@ -22,9 +26,9 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        
-        
-        $projects = Project::latest("created_at")->paginate(8);
+        $newProject = new Project();
+        $projects = Project::/* withTrashed()-> */with("category")->latest("created_at")->paginate(8);
+        $deletedProjects = Project::onlyTrashed()->get();
         
         // $projects = DB::table("project")->get();
 
@@ -33,7 +37,9 @@ class ProjectController extends Controller
         //     "message" => $projects
         // ];
 
-        return view("projects.index")->with(compact("projects"));
+        Debugbar::log(request()->user());
+
+        return view("projects.index")->with(compact("projects", "newProject", "deletedProjects"));
     }
 
     /**
@@ -41,7 +47,15 @@ class ProjectController extends Controller
      */
     public function create()
     {
-        return view("projects.create")->with([ "project" => new Project() ]);
+        if (!Gate::allows("create", $project = new Project())) {
+            return back()->with(["status" => "403 - No autorizado"]);
+            // abort(403);
+        }
+        
+        return view("projects.create")->with([
+            "project" => $project,
+            "categories" => Category::pluck('name', 'id')
+        ]);
     }
     
     /**
@@ -65,12 +79,13 @@ class ProjectController extends Controller
 
         // return dd($request->file("file"));
         
-        
-        $project = new Project($request->only("title", "slug", "description"));
+        $project = new Project($request->only("title", "slug", "description", "category_id"));
 
         $project->image = $request->file("image")->store("images");
             
         $project->save();
+        
+        ProjectSaved::dispatch($project);
         
         return redirect()->route("projects.index")->with("status", __("The project has been created successfully"));
     }
@@ -96,7 +111,9 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
-        return view("projects.edit")->with(compact("project"));
+        $this->authorize('update', $project);
+        $categories = Category::pluck('name', 'id');
+        return view("projects.edit")->with(compact("project", "categories"));
     }
 
     
@@ -107,21 +124,26 @@ class ProjectController extends Controller
     {
         
         // ddd( $request->only("title", "slug", "description", "image") );
+
+        abort_unless(Gate::allows('update', $project), 403);
         
         if ($request->hasFile("image")) {
 
-            Storage::delete($project->image);
+            $project->image && Storage::delete($project->image);
 
-            $project->fill($request->only("title", "slug", "description"));
+            $project->fill($request->only("title", "slug", "description", "category_id"));
             
             $project->image = $request->file("image")->store("images");
             
             $project->save();
+
+            ProjectSaved::dispatch($project);
+           
         } else {
 
             // $project->update(array_filter($request->validated())));
 
-            $project->update($request->only("title", "slug", "description"));
+            $project->update($request->only("title", "slug", "description", "category_id"));
         }
         
 
@@ -135,7 +157,26 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+        $this->authorize("delete", $project);
         $project->delete();
         return redirect()->route("projects.index")->with("status", __("The project has been deleted successfully"));
+    }
+
+
+    public function restore(string $projectSlug)
+    {
+        $project = Project::withTrashed()->whereSlug($projectSlug)->firstOrFail();
+        $this->authorize("restore", $project);
+        $project->restore();
+        return back()->with([ 'status' => __("The project has been restored successfully") ]);
+    }
+
+    public function forceDelete(string $projectSlug)
+    {
+        $project = Project::withTrashed()->whereSlug($projectSlug)->firstOrFail();
+        $this->authorize("forceDelete", $project);
+        Storage::delete($project->image);
+        $project->forceDelete();
+        return back()->with([ 'status' => __("The project has been forcefully deleted successfully") ]);
     }
 }
